+++ .storybook/preview.js
@@ -7,10 +7,6 @@ const theme = buildTheme({
     brandSecondary: '#182848',
   },
   boxes: {
-    overlay: {
-      "background-color": 'rgba(0,0,0, 0.5)',
-      "border-radius": '0px'
-    },
     borderColored: {
       "border-color": '$colors.brandPrimaryClear20',
       "border-width": '2px'