+++ src/molecules/form/component.tsx
@@ -5,27 +5,32 @@ import { ISingleAnyChildProps } from '@kibalabs/core-react';
 import styled from 'styled-components';
 
 import { Box, IBoxTheme, ILoadingSpinnerTheme, LoadingSpinner } from '../../particles';
-import { ThemeType } from '../../util';
+import { useBuiltTheme } from '../../theming';
+import { themeToCss, ThemeType } from '../../util';
 import { defaultMoleculeProps, IMoleculeProps } from '../moleculeProps';
 
-
 export interface IFormTheme extends ThemeType {
-  background: IBoxTheme;
+  backgroundTheme: IBoxTheme;
+  loadingOverlayTheme: IBoxTheme;
   loadingSpinnerTheme: ILoadingSpinnerTheme;
 }
 
-// TODO(krishan711): this should not be relatives when it uses layers
+// TODO(krishan711): this should not be relative when it uses layers
 const StyledForm = styled.form`
   position: relative;
 `;
 
-const LoadingOverlay = styled.div`
+interface ILoadingOverlayProps {
+  $theme: IBoxTheme;
+}
+
+const LoadingOverlay = styled.div<ILoadingOverlayProps>`
+  ${(props: ILoadingOverlayProps): string => themeToCss(props.$theme)};
   width: 100%;
   height: 100%;
   position: absolute;
   top: 0;
   left: 0;
-  background-color: rgba(0, 0, 0, 0.7);
   z-index: 5;
   display: flex;
   flex-direction: column;
@@ -35,6 +40,9 @@ const LoadingOverlay = styled.div`
 
 interface IFormProps extends IMoleculeProps<IFormTheme>, ISingleAnyChildProps {
   isLoading: boolean;
+  backgroundVariant?: string;
+  loadingOverlayVariant?: string;
+  loadingSpinnerVariant?: string;
   onFormSubmitted: () => void;
 }
 
@@ -44,12 +52,15 @@ export const Form = (props: IFormProps): React.ReactElement => {
     props.onFormSubmitted();
   };
 
+  const loadingOverlayTheme = useBuiltTheme('boxes', props.loadingOverlayVariant || 'overlay', props.theme?.loadingOverlayTheme);
+
   // TODO(krishan711): this should use layers
   return (
     <Box
       id={props.id}
       className={getClassName(Form.displayName, props.className)}
-      theme={props.theme?.background}
+      variant={props.backgroundVariant}
+      theme={props.theme?.backgroundTheme}
       isFullWidth={true}
     >
       <StyledForm
@@ -59,8 +70,17 @@ export const Form = (props: IFormProps): React.ReactElement => {
       >
         {props.children}
         { props.isLoading && (
-          <LoadingOverlay id={props.id && `${props.id}-loading-overlay`} className={'form-overlay'}>
-            <LoadingSpinner id={props.id && `${props.id}-loading-spinner`} className={'form-overlay-spinner'} variant='large' />
+          <LoadingOverlay
+            id={props.id && `${props.id}-loading-overlay`}
+            className={'form-overlay'}
+            $theme={loadingOverlayTheme}
+          >
+            <LoadingSpinner
+              id={props.id && `${props.id}-loading-spinner`}
+              className={'form-overlay-spinner'}
+              variant={props.loadingSpinnerVariant}
+              theme={props.theme?.loadingSpinnerTheme}
+            />
           </LoadingOverlay>
         )}
       </StyledForm>