+++ src/util/index.ts
@@ -1,3 +1,4 @@
 export * from './componentUtil';
 export * from './responsiveUtil';
+export * from './variantUtil';
 export * from './theme';